Rename some of the util modules.
