+++ io_scene_nif/io/egm.py
@@ -39,7 +39,7 @@
 
 
 from pyffi.formats.egm import EgmFormat
-from io_scene_nif.utility.nif_logging import NifLog
+from io_scene_nif.utility.util_logging import NifLog
 from io_scene_nif.utility.nif_utils import NifError
