+++ io_scene_nif/modules/animation/animation_export.py
@@ -44,8 +44,8 @@ from pyffi.formats.nif import NifFormat
 
 from io_scene_nif.modules import armature
 from io_scene_nif.utility import nif_utils
-from io_scene_nif.utility.nif_logging import NifLog
-from io_scene_nif.utility.nif_global import NifOp
+from io_scene_nif.utility.util_logging import NifLog
+from io_scene_nif.utility.util_global import NifOp
 
 
 def set_flags_and_timing(kfc, exp_fcurves, start_frame=None, stop_frame=None):
