+++ io_scene_nif/modules/armature/armature_export.py
@@ -37,8 +37,8 @@
 #
 # ***** END LICENSE BLOCK *****
 
-from io_scene_nif.utility.nif_global import NifOp
-from io_scene_nif.utility.nif_logging import NifLog
+from io_scene_nif.utility.util_global import NifOp
+from io_scene_nif.utility.util_logging import NifLog
 
 
 class Armature:
