+++ io_scene_nif/modules/constraint/constraint_export.py
@@ -45,8 +45,8 @@ import mathutils
 from io_scene_nif.modules import collision
 from io_scene_nif.modules.obj.block_registry import block_store
 from io_scene_nif.utility import nif_utils
-from io_scene_nif.utility.nif_logging import NifLog
-from io_scene_nif.utility.nif_global import NifOp
+from io_scene_nif.utility.util_logging import NifLog
+from io_scene_nif.utility.util_global import NifOp
 
 
 class Constraint:
