+++ io_scene_nif/modules/constraint/constraint_import.py
@@ -41,7 +41,7 @@ import mathutils
 from pyffi.formats.nif import NifFormat
 
 from io_scene_nif.modules import collision
-from io_scene_nif.utility.nif_logging import NifLog
+from io_scene_nif.utility.util_logging import NifLog
 
 
 class Constraint:
