+++ io_scene_nif/modules/geometry/morph/morph_export.py
@@ -43,8 +43,8 @@ from pyffi.formats.egm import EgmFormat
 from pyffi.formats.nif import NifFormat
 
 from io_scene_nif.modules.obj.block_registry import block_store
-from io_scene_nif.utility.nif_global import NifOp
-from io_scene_nif.utility.nif_logging import NifLog
+from io_scene_nif.utility.util_global import NifOp
+from io_scene_nif.utility.util_logging import NifLog
 
 
 class MorphHelper:
