+++ io_scene_nif/modules/property/property_export.py
@@ -41,7 +41,7 @@ from pyffi.formats.nif import NifFormat
 
 from io_scene_nif.modules.obj.block_registry import block_store
 from io_scene_nif.modules.property.material.material_export import Material
-from io_scene_nif.utility.nif_global import NifOp
+from io_scene_nif.utility.util_global import NifOp
 
 
 class PropertyHelper:
