+++ io_scene_nif/modules/property/texture/texture_loader.py
@@ -44,8 +44,8 @@ import os.path
 import bpy
 from pyffi.formats.nif import NifFormat
 
-from io_scene_nif.utility.nif_logging import NifLog
-from io_scene_nif.utility.nif_global import NifOp
+from io_scene_nif.utility.util_logging import NifLog
+from io_scene_nif.utility.util_global import NifOp
 from io_scene_nif.modules.property import texture
