+++ io_scene_nif/nif_common.py
@@ -41,8 +41,8 @@ import bpy
 import pyffi
 from pyffi.formats.nif import NifFormat
 
-from io_scene_nif.utility.nif_global import NifOp
-from io_scene_nif.utility.nif_logging import NifLog
+from io_scene_nif.utility.util_global import NifOp
+from io_scene_nif.utility.util_logging import NifLog
 
 
 class NifCommon:
