+++ io_scene_nif/utility/nif_utils.py
@@ -40,7 +40,7 @@ import math
 
 import mathutils
 
-from io_scene_nif.utility.nif_logging import NifLog
+from io_scene_nif.utility.util_logging import NifLog
 
 
 class NifError(Exception):
