+++ io_scene_nif/utility/util_debug.py
@@ -37,8 +37,8 @@
 #
 # ***** END LICENSE BLOCK *****
 
-import sys
 import os
+import sys
 
 CLIENT_PORT = 5678
 REMOTE_PORT = 1234
