+++ io_scene_nif/utility/util_decorator.py
