+++ io_scene_nif/utility/util_global.py
@@ -37,7 +37,7 @@
 #
 # ***** END LICENSE BLOCK *****
 
-from io_scene_nif.utility.nif_logging import NifLog
+from io_scene_nif.utility.util_logging import NifLog
 
 
 class NifOp:
@@ -62,6 +62,14 @@ class NifOp:
 
 
 class NifData:
+
+    data = None
+
+    def __init__(self):
+        pass
+
+
+class KFData:
 
     data = None
