+++ io_scene_nif/utility/util_logging.py
