+++ testframework/integration/modules/geometry/vertex/color/test_vertexcolor.py
@@ -89,7 +89,7 @@ class TestBaseVertexColor(SingleNif):
 
     def b_check_data(self):
         # TODO [geometry] nif file has wrong transform and wrong geometry
-        from io_scene_nif.utility import nif_debug
+        from io_scene_nif.utility import util_debug
         # nif_debug.start_debug()
 
         b_obj = bpy.data.objects[self.b_name]
